html/semantics/document-metadata/the-link-element/link-style-error-01.html

<!DOCTYPE html>
<title>link: error events</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#the-link-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src=/common/get-host-info.sub.js></script>
<div id="log"></div>
<div id="test">
<script>
var t404 = async_test("Should get an error event for a 404 error.")
t404.step(function() {
  var elt = document.createElement("link");
  elt.onerror = t404.step_func(function() {
    assert_true(true, "Got error event for 404 error.")
    t404.step_timeout(function() { t404.done() }, 0);
  })
  elt.onload = t404.unreached_func("load event should not be fired");
  elt.rel = "stylesheet";
  elt.href = "nonexistent_stylesheet.css";
  document.getElementsByTagName("head")[0].appendChild(elt);
})

var tUnsupported = async_test("Should get an error event for an unsupported URL.")
tUnsupported.step(function() {
  var elt = document.createElement("link");
  elt.onerror = tUnsupported.step_func(function() {
    assert_true(true, "Got error event for unsupported URL.")
    tUnsupported.step_timeout(function() { tUnsupported.done() }, 0);
  })
  elt.onload = tUnsupported.unreached_func("load event should not be fired");
  elt.rel = "stylesheet";
  elt.href = "nonexistent:stylesheet.css";
  document.getElementsByTagName("head")[0].appendChild(elt);
})

var tText = async_test("Should get an error event for a text/plain response.")
tText.step(function() {
  var elt = document.createElement("link");
  elt.onerror = tText.step_func(function() {
    assert_true(true, "Got error event for 404 error.")
    tText.done()
  })
  elt.onload = tText.unreached_func("load event should not be fired");
  elt.rel = "stylesheet";
  elt.href = "../../../../../common/css-red.txt";
  document.getElementsByTagName("head")[0].appendChild(elt);
});

[
 ["same-origin", "resources/css.py"],
 ["cross-origin", get_host_info().HTTP_REMOTE_ORIGIN + "/html/semantics/document-metadata/the-link-element/resources/css.py"]
].forEach(originType => {
  ["no Content-Type", "broken Content-Type"].forEach(contentType => {
    ["no nosniff", "nosniff"].forEach(nosniff => {
      async_test(t => {
        const l = document.createElement("link");
        t.add_cleanup(() => l.remove());
        if (nosniff === "nosniff") {
          l.onerror = t.step_func_done();
          l.onload = t.unreached_func("error event should have fired");
        } else {
          l.onload = t.step_func_done();
          l.onerror = t.unreached_func("load event should have fired");
        }
        l.rel = "stylesheet";
        let query = [];
        if (contentType === "broken Content-Type") {
          query.push("attempt_content_type");
        }
        if (nosniff === "nosniff") {
          query.push("nosniff");
        }
        let stringQuery = "";
        query.forEach(val => {
          if (stringQuery === "") {
            stringQuery += "?" + val;
          } else {
            stringQuery += "&" + val;
          }
        });
        l.href = originType[1] + stringQuery;
        document.head.appendChild(l);
      }, "Stylesheet with " + contentType + ", " + originType[0] + ", and " + nosniff);
    });
  });
});
</script>
