LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-link-element/link-style-error-01-expected.txt

CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/common/css-red.txt' because non CSS MIME types are not allowed in strict mode.

PASS Should get an error event for a 404 error. 
PASS Should get an error event for an unsupported URL. 
FAIL Should get an error event for a text/plain response. assert_unreached: load event should not be fired Reached unreachable code
